Begin calendar plugin; improve event metadata
Instead of separate start and end metadata fields, allow specifying the
event in a single field (with the syntax "Event: YYYY-MM-DD HH:MM p to
HH:MM p") which is validated and unpacked into datetime.datetime()
instances by a Pelican plugin.  This should be easier for people to
author, as the date only need be specified once and 12-hour time is
expected.